Implement user creation and login API routes
This commits adds new services, controllers, and models as necessary to
implement the user creation and login API routes. Their documentation is
included in Docs/ and their associated tests are included in Tests/. It
also includes miscellaneous fixes, like a versioning fix for the
dependencies and some database schema adjustments necessary for the new
feature.