Implement basic user post CRUD functionality
This commit adds basic functionality for user posts. Currently, a user
receives a JWT on successful sign-in. With this JWT, posts can be made
from an endpoint, as well as updated or deleted. Posts can be queried
with a variety of parameters without the need for a JWT. Documentation
for this is in `Docs/API.md`. Additionally, there are small tweaks to
some classnames, members, etc. Some of these are actually breaking
changes, but nobody has used the backend APIs yet. Woo!