Implement petitions
This commit implements petitions as a sub-type of post. Despite sharing
a lot of overlap with posts, they have separate endpoints, and
operations to each set of end points will generally only be able to
interact with the associated type. Aside from adding petitions,
documentation, and tests, this commit also makes minor adjustments to
the behavior of posts to clearly delineate them from petitions.